decluttr-app/backend/pyproject.toml
[tool.poetry]
name = "decluttr-app/backend"
version = "0.1.0"
description = ""
authors = ["Isabelle <isisdesade@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11,<3.12"
fastapi = "^0.104.1"
uvicorn = { extras = ["standard"], version = "^0.23.2" }
llama-index = "^0.9.15.post2"
pypdf = "^3.17.0"
python-dotenv = "^1.0.0"
trulens-eval = "^0.19.2"
google-generativeai = ">=0.3.0"
supabase = "^2.3.0"
mangum = "^0.17.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.27.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
